requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
numpy>=1.23.0
sentence-transformers>=2.7.0
torch>=2.0.0
